app/requirements.txt
docopt==0.6.2
ducktoolkit==1.0.3
factordb-pycli==1.3.0
Flask==2.2.5
librosa==0.9.2
matplotlib==3.5.2
numpy==1.22
oletools
opencv_python==4.6.0.66
Pillow==9.3.0
pyOpenSSL==21.0.0
pwnlib==1.2.1
pwntools==4.8.0
pycryptodome==3.15.0
pyopenssl
pydub==0.25.1
py-midi==2.0.1
pypcapkit==0.16.2
pyUnicodeSteganography==0.0.1
r2pipe==1.7.3
scapy==2.4.5
scipy
stegano
stegoveritas==1.9
stegpy==0.0.1
stringcheese==2.1
tinyscript==1.26.11
Werkzeug==2.2.3